requirements.txt
Django>=5.2,<5.3
mysqlclient>=2.2,<2.3
python-dotenv>=1.0,<2.0
